account/api/serializers.py
from rest_framework import serializers
from account.models import Account


class RegistrationSerializer(serializers.ModelSerializer):
    #account model doesnt has 2 password field so we need to create one 
    password2 = serializers.CharField(
        style={'input_type': 'password'}, write_only=True)

    class Meta:
        model = Account
        fields = ['email', 'username', 'password', 'password2']
        #for security password should be hidden
        extra_kwargs = {
            'password':  {'write_only': True}
        }

    def save(self):
        account = Account(
            email=self.validated_data['email'],
            username=self.validated_data['username'],
        )
        password = self.validated_data['password']
        password2 = self.validated_data['password2']

        if password != password2:
            raise serializers.ValidationError(
                {'password': 'passwords must match.'})
        account.set_password(password)
        account.save()
        return account
