requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.2
django-heroku==0.3.1
djangorestframework==3.12.1
django-cors-headers==3.6.0
filelock==3.0.12
gunicorn==20.0.4
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
mypy==0.790
mypy-extensions==0.4.3
Pillow==8.0.0
psycopg2==2.8.6
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
virtualenv==20.0.35
whitenoise==5.2.0
wrapt==1.12.1
